Added repositories, services, dto, mappers and comparator

+++ src/main/java/com/cursor/lobo/moviereview/entity/Movie.java
@@ -48,7 +48,7 @@ public class Movie {
     private String shortDescription;
 
     @Column(name = "rateValue")
-    private int rateValue;
+    private Double rateValue;
 
     @OneToOne(mappedBy = "movie")
     private Rate rate;
